+++ scripts/specificSummons.js
@@ -30,7 +30,7 @@ export async function getSpecificSummonDetails(
     targetTokenUUID: null,
     tokenWidth: 1,
     tokenHeight: 1,
-  }
+  },
 ) {
   if (isSummonSourceDisabled(uuid)) {
     return null;
@@ -71,6 +71,7 @@ const getSummonHandlers = () => ({
   [SOURCES.MISC.AVENGING_WILDWOOD]: handlers.misc.handleAvengingWildwood,
   [SOURCES.MISC.CALL_URSINE_ALLY]: handlers.misc.handleCallUrsineAlly,
   [SOURCES.MISC.DUPLICATE_FOE]: handlers.misc.handleDuplicateFoe,
+  [SOURCES.MISC.DRAGON_TURRET]: handlers.misc.handleDragonTurret,
   [SOURCES.MISC.FLOATING_FLAME]: handlers.misc.handleFloatingFlame,
   [SOURCES.MISC.LIGHT]: handlers.misc.handleLight,
   [SOURCES.MISC.TELEKINETIC_HAND]: handlers.misc.handleTelekineticHand,
@@ -246,7 +247,7 @@ const handlers = {
           },
           itemsToAdd: [
             EFFECTS.RULE_EFFECT(
-              getAvengingWildwoodStrikeRuleElements({ rank: data.rank })
+              getAvengingWildwoodStrikeRuleElements({ rank: data.rank }),
             ),
           ],
         },
@@ -270,7 +271,7 @@ const handlers = {
       if (token) {
         if (token?.actor?.level > maxLevel) {
           errorNotification(
-            "pf2e-summons-assistant.notification.duplicate-foe.too-high"
+            "pf2e-summons-assistant.notification.duplicate-foe.too-high",
           );
           return null;
         }
@@ -278,7 +279,7 @@ const handlers = {
         const info = await getFoeInfo(token, data.rank);
         const isFail = await foundry.applications.api.DialogV2.confirm({
           content: game.i18n.localize(
-            "pf2e-summons-assistant.dialog.duplicate-foe"
+            "pf2e-summons-assistant.dialog.duplicate-foe",
           ),
           rejectClose: false,
         });
@@ -297,6 +298,15 @@ const handlers = {
         ];
       }
       return null;
+    },
+
+    handleDragonTurret: async () => {
+      return [
+        {
+          specific_uuids: [CREATURES.DRAGON_TURRET],
+          itemsToAdd: [EFFECTS.RULE_EFFECT([RULE_ELEMENTS.SPELL_DC_FLAG])],
+        },
+      ];
     },
 
     handleFloatingFlame: async (data) => {
@@ -343,7 +353,7 @@ const handlers = {
     handleTelekineticHand: async (data) => {
       const isInvisible = await foundry.applications.api.DialogV2.confirm({
         content: game.i18n.localize(
-          "pf2e-summons-assistant.dialog.telekinetic-hand"
+          "pf2e-summons-assistant.dialog.telekinetic-hand",
         ),
         rejectClose: false,
       });
@@ -390,7 +400,7 @@ const handlers = {
             },
             label: {
               text: game.i18n.localize(
-                "pf2e-summons-assistant.display-text.wooden-double.place-double"
+                "pf2e-summons-assistant.display-text.wooden-double.place-double",
               ),
             },
             ...(data.position
@@ -437,7 +447,7 @@ const handlers = {
       const type = await foundry.applications.api.DialogV2.wait({
         window: { title: "Wall of Fire" },
         content: await TextEditor.enrichHTML(
-          `<p>${game.i18n.localize("pf2e-summons-assistant.dialog.choose-type-of")} @UUID[${SOURCES.WALL.WALL_OF_FIRE}]</p>`
+          `<p>${game.i18n.localize("pf2e-summons-assistant.dialog.choose-type-of")} @UUID[${SOURCES.WALL.WALL_OF_FIRE}]</p>`,
         ),
         // This example does not use i18n strings for the button labels,
         // but they are automatically localized.
@@ -478,7 +488,7 @@ const handlers = {
                 crosshairParameters: {
                   label: {
                     text: game.i18n.localize(
-                      "pf2e-summons-assistant.display-text.wall.start-point"
+                      "pf2e-summons-assistant.display-text.wall.start-point",
                     ),
                   },
                 },