+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Unreleased
 
+## 1.21.0
+
+- **New**
+  - Added support for `Dragon Turret` (art by @Chasarooni)
+- **Updated**
+  - Handled situation where a token has no actor breaking `Wall of Fire` (🐛 @zionhian)
+  - Added `Floating Flame` and `Avenging Wildwood` to summon's who's DCs are linjed to their summoner
+
 ## 1.20.3
 
 - Push data from `0.20.2`